requirements.txt
# RAG Chatbot for Agentic AI eBook
# Core dependencies with pinned versions for reproducibility

python-dotenv>=1.0.0
streamlit>=1.20.0
pdfplumber>=0.10.3
sentence-transformers>=5.0.0
torch>=2.0.0
pinecone>=3.0.0
langgraph>=0.0.40
tqdm>=4.66.1
tiktoken>=0.5.2
openai>=1.0.0
groq>=0.4.0
tf-keras

# NOTE: For CPU-only machines, you can install torch CPU version:
# pip install torch --index-url https://download.pytorch.org/whl/cpu
# This will be smaller and faster to install.
